Remove cross-endpoint text relay bridging
