+++ internal/relay/adaptor/anthropic/messages_passthrough_test.go
@@ -144,43 +144,6 @@ func TestRelayMessagesStreamResponseSupportsLargeAnthropicDataLine(t *testing.T)
 	}
 }
 
-func TestRelayMessagesStreamAsChatResponse(t *testing.T) {
-	ctx, recorder := newAnthropicPassthroughTestContext()
-	resp := &http.Response{
-		StatusCode: http.StatusOK,
-		Header: http.Header{
-			"Content-Type": []string{"text/event-stream"},
-			"X-Upstream":   []string{"ok"},
-		},
-		Body: io.NopCloser(strings.NewReader(
-			"event: message_start\n" +
-				"data: {\"type\":\"message_start\",\"message\":{\"id\":\"msg_1\",\"model\":\"claude-opus-4-6\",\"usage\":{\"input_tokens\":11,\"output_tokens\":0}}}\n\n" +
-				"event: content_block_delta\n" +
-				"data: {\"type\":\"content_block_delta\",\"delta\":{\"type\":\"text_delta\",\"text\":\"hello \"}}\n\n" +
-				"event: content_block_delta\n" +
-				"data: {\"type\":\"content_block_delta\",\"delta\":{\"type\":\"text_delta\",\"text\":\"world\"}}\n\n" +
-				"event: message_delta\n" +
-				"data: {\"type\":\"message_delta\",\"delta\":{\"stop_reason\":\"end_turn\"},\"usage\":{\"output_tokens\":7}}\n\n" +
-				"data: [DONE]\n",
-		)),
-	}
-
-	usage, relayErr := relayMessagesStreamAsChatResponse(ctx, resp, 11, "claude-opus-4-6")
-	if relayErr != nil {
-		t.Fatalf("relayMessagesStreamAsChatResponse returned error: %+v", relayErr)
-	}
-	if usage == nil || usage.PromptTokens != 11 || usage.CompletionTokens != 7 || usage.TotalTokens != 18 {
-		t.Fatalf("unexpected usage: %#v", usage)
-	}
-	body := recorder.Body.String()
-	if !strings.Contains(body, `"object":"chat.completion"`) || !strings.Contains(body, `"content":"hello world"`) {
-		t.Fatalf("unexpected bridged chat payload: %s", body)
-	}
-	if recorder.Header().Get("X-Upstream") != "ok" {
-		t.Fatalf("expected upstream header to be copied, got %q", recorder.Header().Get("X-Upstream"))
-	}
-}
-
 func TestRelayMessagesResponseSkipsUpstreamCORSHeaders(t *testing.T) {
 	ctx, recorder := newAnthropicPassthroughTestContext()
 	recorder.Header().Set("Access-Control-Allow-Origin", "http://localhost:3020")
@@ -211,37 +174,5 @@ func TestRelayMessagesResponseSkipsUpstreamCORSHeaders(t *testing.T) {
 	}
 	if recorder.Header().Get("X-Upstream") != "ok" {
 		t.Fatalf("expected non-CORS upstream headers to remain copied, got %q", recorder.Header().Get("X-Upstream"))
-	}
-}
-
-func TestRelayMessagesStreamAsChatResponseSkipsUpstreamCORSHeaders(t *testing.T) {
-	ctx, recorder := newAnthropicPassthroughTestContext()
-	recorder.Header().Set("Access-Control-Allow-Origin", "http://localhost:3020")
-	recorder.Header().Set("Access-Control-Allow-Credentials", "true")
-	resp := &http.Response{
-		StatusCode: http.StatusOK,
-		Header: http.Header{
-			"Content-Type":                     []string{"text/event-stream"},
-			"Access-Control-Allow-Origin":      []string{"*"},
-			"Access-Control-Allow-Credentials": []string{"true"},
-		},
-		Body: io.NopCloser(strings.NewReader(
-			"event: message_start\n" +
-				"data: {\"type\":\"message_start\",\"message\":{\"id\":\"msg_1\",\"model\":\"claude-opus-4-6\",\"usage\":{\"input_tokens\":11,\"output_tokens\":0}}}\n\n" +
-				"event: content_block_delta\n" +
-				"data: {\"type\":\"content_block_delta\",\"delta\":{\"type\":\"text_delta\",\"text\":\"hello\"}}\n\n" +
-				"event: message_delta\n" +
-				"data: {\"type\":\"message_delta\",\"delta\":{\"stop_reason\":\"end_turn\"},\"usage\":{\"output_tokens\":7}}\n\n" +
-				"data: [DONE]\n",
-		)),
-	}
-
-	_, relayErr := relayMessagesStreamAsChatResponse(ctx, resp, 11, "claude-opus-4-6")
-	if relayErr != nil {
-		t.Fatalf("relayMessagesStreamAsChatResponse returned error: %+v", relayErr)
-	}
-	allowOriginValues := recorder.Header().Values("Access-Control-Allow-Origin")
-	if len(allowOriginValues) != 1 || allowOriginValues[0] != "http://localhost:3020" {
-		t.Fatalf("expected CORS allow-origin header to remain middleware value, got %#v", allowOriginValues)
 	}
 }
