+++ internal/relay/controller/text_bridge_test.go
@@ -198,7 +198,7 @@ func TestResolveChannelTextUpstreamNoModelConfigsOpenAIMessagesRejected(t *testi
 	}
 }
 
-func TestConvertTextRequestForUpstreamToResponses(t *testing.T) {
+func TestConvertTextRequestForUpstreamRejectsChatToResponses(t *testing.T) {
 	req := &relaymodel.GeneralOpenAIRequest{
 		Model: "gpt-4.1",
 		Messages: []relaymodel.Message{{
@@ -208,174 +208,52 @@ func TestConvertTextRequestForUpstreamToResponses(t *testing.T) {
 		MaxTokens: 128,
 	}
 
-	converted, err := convertTextRequestForUpstream(req, relaymode.ChatCompletions, relaymode.Responses)
-	if err != nil {
-		t.Fatalf("convertTextRequestForUpstream returned error: %v", err)
-	}
-	if len(converted.Messages) != 0 {
-		t.Fatalf("converted.Messages = %#v, want empty", converted.Messages)
-	}
-	if converted.Input == nil {
-		t.Fatalf("converted.Input = nil, want messages copied into input")
-	}
-	if converted.MaxOutputTokens == nil || *converted.MaxOutputTokens != 128 {
-		t.Fatalf("converted.MaxOutputTokens = %#v, want 128", converted.MaxOutputTokens)
+	if _, err := convertTextRequestForUpstream(req, relaymode.ChatCompletions, relaymode.Responses); err == nil {
+		t.Fatalf("convertTextRequestForUpstream returned nil error, want endpoint-conversion error")
 	}
 }
 
-func TestConvertTextRequestForUpstreamToResponsesWithImageContent(t *testing.T) {
-	req := &relaymodel.GeneralOpenAIRequest{
-		Model: "gpt-5.4",
-		Messages: []relaymodel.Message{{
-			Role: "user",
-			Content: []any{
-				map[string]any{
-					"type": "text",
-					"text": "describe this image",
-				},
-				map[string]any{
-					"type": "image_url",
-					"image_url": map[string]any{
-						"url":    "https://example.com/a.png",
-						"detail": "high",
-					},
-				},
-			},
-		}},
-	}
-
-	converted, err := convertTextRequestForUpstream(req, relaymode.ChatCompletions, relaymode.Responses)
-	if err != nil {
-		t.Fatalf("convertTextRequestForUpstream returned error: %v", err)
-	}
-	inputList, ok := converted.Input.([]any)
-	if !ok || len(inputList) != 1 {
-		t.Fatalf("converted.Input = %#v, want one-item input list", converted.Input)
-	}
-	first, ok := inputList[0].(map[string]any)
-	if !ok {
-		t.Fatalf("converted.Input[0] = %#v, want map", inputList[0])
-	}
-	contentList, ok := first["content"].([]any)
-	if !ok || len(contentList) != 2 {
-		t.Fatalf("converted.Input[0].content = %#v, want two content blocks", first["content"])
-	}
-	textPart, ok := contentList[0].(map[string]any)
-	if !ok || textPart["type"] != "input_text" || textPart["text"] != "describe this image" {
-		t.Fatalf("contentList[0] = %#v, want input_text", contentList[0])
-	}
-	imagePart, ok := contentList[1].(map[string]any)
-	if !ok || imagePart["type"] != "input_image" || imagePart["image_url"] != "https://example.com/a.png" || imagePart["detail"] != "high" {
-		t.Fatalf("contentList[1] = %#v, want input_image", contentList[1])
-	}
-}
-
-func TestConvertTextRequestForUpstreamToResponsesWithImageURLStringContent(t *testing.T) {
-	req := &relaymodel.GeneralOpenAIRequest{
-		Model: "gpt-5.4",
-		Messages: []relaymodel.Message{{
-			Role: "user",
-			Content: []any{
-				map[string]any{
-					"type":      "image_url",
-					"image_url": "https://example.com/raw.png",
-				},
-			},
-		}},
-	}
-
-	converted, err := convertTextRequestForUpstream(req, relaymode.ChatCompletions, relaymode.Responses)
-	if err != nil {
-		t.Fatalf("convertTextRequestForUpstream returned error: %v", err)
-	}
-	inputList, ok := converted.Input.([]any)
-	if !ok || len(inputList) != 1 {
-		t.Fatalf("converted.Input = %#v, want one-item input list", converted.Input)
-	}
-	first, ok := inputList[0].(map[string]any)
-	if !ok {
-		t.Fatalf("converted.Input[0] = %#v, want map", inputList[0])
-	}
-	contentList, ok := first["content"].([]any)
-	if !ok || len(contentList) != 1 {
-		t.Fatalf("converted.Input[0].content = %#v, want one content block", first["content"])
-	}
-	imagePart, ok := contentList[0].(map[string]any)
-	if !ok || imagePart["type"] != "input_image" || imagePart["image_url"] != "https://example.com/raw.png" {
-		t.Fatalf("contentList[0] = %#v, want input_image with image_url", contentList[0])
-	}
-}
-
-func TestConvertTextRequestForUpstreamToChat(t *testing.T) {
+func TestConvertTextRequestForUpstreamRejectsResponsesToChat(t *testing.T) {
 	req := &relaymodel.GeneralOpenAIRequest{
 		Model:           "gpt-4.1",
 		Input:           "hello",
 		MaxOutputTokens: func() *int { value := 256; return &value }(),
 	}
 
-	converted, err := convertTextRequestForUpstream(req, relaymode.Responses, relaymode.ChatCompletions)
-	if err != nil {
-		t.Fatalf("convertTextRequestForUpstream returned error: %v", err)
-	}
-	if len(converted.Messages) != 1 || converted.Messages[0].StringContent() != "hello" {
-		t.Fatalf("converted.Messages = %#v, want single user message", converted.Messages)
-	}
-	if converted.Input != nil {
-		t.Fatalf("converted.Input = %#v, want nil", converted.Input)
-	}
-	if converted.MaxTokens != 256 {
-		t.Fatalf("converted.MaxTokens = %d, want 256", converted.MaxTokens)
+	if _, err := convertTextRequestForUpstream(req, relaymode.Responses, relaymode.ChatCompletions); err == nil {
+		t.Fatalf("convertTextRequestForUpstream returned nil error, want endpoint-conversion error")
 	}
 }
 
-func TestConvertTextRequestForUpstreamToMessagesFromInput(t *testing.T) {
+func TestConvertTextRequestForUpstreamRejectsResponsesToMessages(t *testing.T) {
 	req := &relaymodel.GeneralOpenAIRequest{
 		Model:           "claude-sonnet-4-6",
-		Input:           "hello from responses input",
-		MaxOutputTokens: func() *int { value := 320; return &value }(),
-	}
-
-	converted, err := convertTextRequestForUpstream(req, relaymode.Responses, relaymode.Messages)
-	if err != nil {
-		t.Fatalf("convertTextRequestForUpstream returned error: %v", err)
-	}
-	if converted.Input != nil {
-		t.Fatalf("converted.Input = %#v, want nil", converted.Input)
-	}
-	if len(converted.Messages) != 1 || converted.Messages[0].StringContent() != "hello from responses input" {
-		t.Fatalf("converted.Messages = %#v, want single converted user message", converted.Messages)
-	}
-	if converted.MaxTokens != 320 {
-		t.Fatalf("converted.MaxTokens = %d, want 320", converted.MaxTokens)
-	}
-}
-
-func TestConvertTextRequestForUpstreamToMessagesPreservesResponsesInstructions(t *testing.T) {
-	req := &relaymodel.GeneralOpenAIRequest{
-		Model:           "claude-opus-4.6",
 		Input:           "hello from responses input",
 		Instructions:    "reply in haiku form",
 		MaxOutputTokens: func() *int { value := 320; return &value }(),
 	}
 
-	converted, err := convertTextRequestForUpstream(req, relaymode.Responses, relaymode.Messages)
+	if _, err := convertTextRequestForUpstream(req, relaymode.Responses, relaymode.Messages); err == nil {
+		t.Fatalf("convertTextRequestForUpstream returned nil error, want endpoint-conversion error")
+	}
+}
+
+func TestConvertTextRequestForUpstreamSameModePassesThrough(t *testing.T) {
+	req := &relaymodel.GeneralOpenAIRequest{
+		Model: "gpt-4.1",
+		Messages: []relaymodel.Message{{
+			Role:    "user",
+			Content: "hello",
+		}},
+		MaxTokens: 128,
+	}
+
+	converted, err := convertTextRequestForUpstream(req, relaymode.ChatCompletions, relaymode.ChatCompletions)
 	if err != nil {
 		t.Fatalf("convertTextRequestForUpstream returned error: %v", err)
 	}
-	if len(converted.Messages) != 2 {
-		t.Fatalf("len(converted.Messages) = %d, want 2", len(converted.Messages))
-	}
-	if converted.Messages[0].Role != "system" || converted.Messages[0].StringContent() != "reply in haiku form" {
-		t.Fatalf("converted.Messages[0] = %#v, want system instructions message", converted.Messages[0])
-	}
-	if converted.Messages[1].Role != "user" || converted.Messages[1].StringContent() != "hello from responses input" {
-		t.Fatalf("converted.Messages[1] = %#v, want converted user input message", converted.Messages[1])
-	}
-	if converted.Input != nil {
-		t.Fatalf("converted.Input = %#v, want nil", converted.Input)
-	}
-	if converted.MaxTokens != 320 {
-		t.Fatalf("converted.MaxTokens = %d, want 320", converted.MaxTokens)
+	if len(converted.Messages) != 1 || converted.Messages[0].StringContent() != "hello" {
+		t.Fatalf("converted.Messages = %#v, want original chat message", converted.Messages)
 	}
 }
 
@@ -395,138 +273,5 @@ func TestNormalizeMessagesRequestBodyUpdatesModel(t *testing.T) {
 	}
 	if payload["stream"] != true {
 		t.Fatalf("payload.stream = %#v, want true", payload["stream"])
-	}
-}
-
-func TestNormalizeRequestBodyForResponsesConvertsToolsChoiceAndFormat(t *testing.T) {
-	raw := []byte(`{
-		"model":"gpt-5.4",
-		"max_completion_tokens":128,
-		"response_format":{
-			"type":"json_schema",
-			"json_schema":{
-				"name":"calendar_event",
-				"schema":{"type":"object","properties":{"title":{"type":"string"}},"required":["title"]},
-				"strict":true
-			}
-		},
-		"tools":[
-			{
-				"type":"function",
-				"function":{
-					"name":"lookup_weather",
-					"description":"lookup weather",
-					"parameters":{"type":"object","properties":{"city":{"type":"string"}},"required":["city"]},
-					"strict":true
-				}
-			}
-		],
-		"tool_choice":{
-			"type":"function",
-			"function":{"name":"lookup_weather"}
-		},
-		"n":1
-	}`)
-	normalized, err := normalizeRequestBodyForResponses(raw)
-	if err != nil {
-		t.Fatalf("normalizeRequestBodyForResponses returned error: %v", err)
-	}
-	payload := map[string]any{}
-	if err := json.Unmarshal(normalized, &payload); err != nil {
-		t.Fatalf("json.Unmarshal normalized body returned error: %v", err)
-	}
-	if _, exists := payload["response_format"]; exists {
-		t.Fatalf("payload.response_format should be removed, got %#v", payload["response_format"])
-	}
-	if _, exists := payload["n"]; exists {
-		t.Fatalf("payload.n should be removed, got %#v", payload["n"])
-	}
-	if payload["max_output_tokens"] != float64(128) {
-		t.Fatalf("payload.max_output_tokens = %#v, want 128", payload["max_output_tokens"])
-	}
-	if _, exists := payload["max_completion_tokens"]; exists {
-		t.Fatalf("payload.max_completion_tokens should be removed, got %#v", payload["max_completion_tokens"])
-	}
-	text, ok := payload["text"].(map[string]any)
-	if !ok {
-		t.Fatalf("payload.text = %#v, want map", payload["text"])
-	}
-	format, ok := text["format"].(map[string]any)
-	if !ok {
-		t.Fatalf("payload.text.format = %#v, want map", text["format"])
-	}
-	if format["type"] != "json_schema" || format["name"] != "calendar_event" || format["strict"] != true {
-		t.Fatalf("payload.text.format = %#v, want json_schema with name/strict", format)
-	}
-	tools, ok := payload["tools"].([]any)
-	if !ok || len(tools) != 1 {
-		t.Fatalf("payload.tools = %#v, want one tool", payload["tools"])
-	}
-	tool, ok := tools[0].(map[string]any)
-	if !ok {
-		t.Fatalf("payload.tools[0] = %#v, want map", tools[0])
-	}
-	if tool["type"] != "function" || tool["name"] != "lookup_weather" {
-		t.Fatalf("payload.tools[0] = %#v, want flattened function tool", tool)
-	}
-	if _, exists := tool["function"]; exists {
-		t.Fatalf("payload.tools[0].function should be removed, got %#v", tool["function"])
-	}
-	toolChoice, ok := payload["tool_choice"].(map[string]any)
-	if !ok {
-		t.Fatalf("payload.tool_choice = %#v, want map", payload["tool_choice"])
-	}
-	if toolChoice["type"] != "function" || toolChoice["name"] != "lookup_weather" {
-		t.Fatalf("payload.tool_choice = %#v, want flattened function choice", toolChoice)
-	}
-}
-
-func TestNormalizeRequestBodyForResponsesConvertsLegacyFunctionsAndFunctionCall(t *testing.T) {
-	raw := []byte(`{
-		"model":"gpt-5.4",
-		"functions":[
-			{
-				"name":"legacy_lookup",
-				"description":"legacy",
-				"parameters":{"type":"object","properties":{"city":{"type":"string"}},"required":["city"]}
-			}
-		],
-		"function_call":{"name":"legacy_lookup"},
-		"max_tokens":64
-	}`)
-	normalized, err := normalizeRequestBodyForResponses(raw)
-	if err != nil {
-		t.Fatalf("normalizeRequestBodyForResponses returned error: %v", err)
-	}
-	payload := map[string]any{}
-	if err := json.Unmarshal(normalized, &payload); err != nil {
-		t.Fatalf("json.Unmarshal normalized body returned error: %v", err)
-	}
-	if _, exists := payload["functions"]; exists {
-		t.Fatalf("payload.functions should be removed, got %#v", payload["functions"])
-	}
-	if _, exists := payload["function_call"]; exists {
-		t.Fatalf("payload.function_call should be removed, got %#v", payload["function_call"])
-	}
-	if payload["max_output_tokens"] != float64(64) {
-		t.Fatalf("payload.max_output_tokens = %#v, want 64", payload["max_output_tokens"])
-	}
-	tools, ok := payload["tools"].([]any)
-	if !ok || len(tools) != 1 {
-		t.Fatalf("payload.tools = %#v, want one legacy-converted tool", payload["tools"])
-	}
-	tool, ok := tools[0].(map[string]any)
-	if !ok {
-		t.Fatalf("payload.tools[0] = %#v, want map", tools[0])
-	}
-	if tool["type"] != "function" || tool["name"] != "legacy_lookup" {
-		t.Fatalf("payload.tools[0] = %#v, want converted legacy function", tool)
-	}
-	toolChoice, ok := payload["tool_choice"].(map[string]any)
-	if !ok {
-		t.Fatalf("payload.tool_choice = %#v, want map", payload["tool_choice"])
-	}
-	if toolChoice["type"] != "function" || toolChoice["name"] != "legacy_lookup" {
-		t.Fatalf("payload.tool_choice = %#v, want legacy function_call conversion", toolChoice)
 	}
 }
